% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profileCI-logLikFn.R
\name{logLikFn}
\alias{logLikFn}
\alias{logLikFn.glm}
\title{Calculate Log-Likelihood}
\usage{
logLikFn(object, pars, ...)

\method{logLikFn}{glm}(object, pars, ...)
}
\arguments{
\item{object}{A fitted model object.}

\item{pars}{A numeric vector of parameters of the model.}

\item{...}{Further arguments.}
}
\value{
A numberic scalar. The value of the log-likelihood function for the
fitted model object \code{object} for parameter values \code{pars}.
}
\description{
This is a generic function for calculating a log-likelihood for an object
for input parameter values.
}
\details{
This generic function has been created to enable a function that
calculates the log-likelihood for a parametric model at a given set of
parameter values in \code{pars} to be available to the function \code{\link{profileCI}}.
}
