% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profileCI-package.R
\docType{package}
\name{profileCI-package}
\alias{profileCI-package}
\title{profileCI: Profiling a Log-Likelihood to Calculate a Confidence Interval}
\description{
Describe
}
\details{
The main function is \code{\link{profileCI}}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://paulnorthrop.github.io/profileCI/}
  \item \url{https://github.com/paulnorthrop/profileCI}
  \item Report bugs at \url{https://github.com/paulnorthrop/profileCI/issues}
}

}
\author{
\strong{Maintainer}: Paul J. Northrop \email{p.northrop@ucl.ac.uk} [copyright holder]

}
