% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profileCI-logLikFn.R
\name{logLikFn}
\alias{logLikFn}
\alias{logLikFn.glm}
\title{Calculate Log-Likelihood}
\usage{
logLikFn(object, pars, ...)

\method{logLikFn}{glm}(object, pars, ...)
}
\arguments{
\item{object}{A fitted model object.}

\item{pars}{A numeric vector of parameters of the model.}

\item{...}{Further arguments.}
}
\description{
This is a generic function for calculating a log-likelihood for an object
for input parameter values.
}
