% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profileCI-internal.R
\name{profileCI-internal}
\alias{profileCI-internal}
\alias{profile_ci}
\alias{faster_profile_ci}
\alias{lagrangianInterpolation}
\alias{initial_mvn}
\title{Internal profileCI functions}
\usage{
profile_ci(
  object,
  negated_loglik_fn,
  which = 1,
  level,
  mle,
  inc,
  epsilon,
  optim_args,
  mult,
  flat,
  ...
)

faster_profile_ci(
  object,
  negated_loglik_fn,
  which = 1,
  which_name,
  level,
  mle,
  ci_sym_mat,
  inc,
  epsilon,
  optim_args,
  mult,
  flat,
  ...
)

lagrangianInterpolation(x0, y0)

initial_mvn(object, x, x_value)
}
\description{
Internal profileCI functions
}
\details{
These functions are not intended to be called by the user.
}
\keyword{internal}
