% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profileCI-internal.R
\name{profileCI-internal}
\alias{profileCI-internal}
\alias{profile_ci}
\alias{faster_profile_ci}
\alias{gev_profile_init}
\alias{lagrangianInterpolation}
\title{Internal profileCI functions}
\usage{
profile_ci(negated_loglik_fn, which = 1, level, mle, inc, epsilon, ...)

faster_profile_ci(
  negated_loglik_fn,
  which = 1,
  which_name,
  level,
  mle,
  ci_sym_mat,
  inc,
  epsilon,
  ...
)

gev_profile_init(data, mu, sigma, xi)

lagrangianInterpolation(x0, y0)
}
\description{
Internal profileCI functions
}
\details{
These functions are not intended to be called by the user.
}
\keyword{internal}
